requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.16.0
transformers>=4.30.0
accelerate>=0.20.3
peft>=0.3.0
datasets>=2.10.1
Pillow>=9.0.0
onnxruntime>=1.15.1
tqdm>=4.64.0
PyYAML>=6.0
scipy>=1.7.0
scikit-learn>=1.0
scikit-image>=0.19.0
opencv-python>=4.5.1
opencv-python-headless>=4.5.1
wandb>=0.15.0
gradio>=3.0.0
git+